THE-505: Extract S3 listing handlers
